# backend/openapi.yaml
openapi: 3.0.3
info:
  title: HR / E-Learning API
  version: "1.0.0"
  description: |
    OpenAPI spec generated from your NestJS controllers & DTOs (User, Auth, Admin, Feedback, AuditLog, Backup).
    Sensitive fields (passwordHash, mfaSecret, mfaBackupCodes) are intentionally excluded from public response schemas.
servers:
  - url: http://localhost:3786
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Student, Instructor, Admin]
        profileImage:
          type: string
    PublicUserDto:
      type: object
      properties:
        _id:
          type: string
          example: "64b8f7a0c2f4e9b3f1a2d3c4"
        name:
          type: string
          example: "Alice Example"
        email:
          type: string
          format: email
          example: "alice@example.com"
        role:
          type: string
          example: "Student"
        isEmailVerified:
          type: boolean
          example: false
        unreadNotificationCount:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedUserDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicUserDto'
        total:
          type: integer
          example: 234
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 12
        limit:
          type: integer
          example: 20

    # -------------------------
    # Auth
    # -------------------------
...
    CreateFeedbackDto:
      type: object
      properties:
        message:
          type: string
          maxLength: 4000
        contactEmail:
          type: string
          format: email
        category:
          type: string
          example: "general"
      required: [message]
    PublicFeedbackDto:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
          nullable: true
        message:
          type: string
        contactEmail:
          type: string
          nullable: true
        category:
          type: string
          example: general
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # -------------------------
    # Audit Log
    # -------------------------
    AuditEvent:
      type: string
      enum:
        - EMAIL_VERIFIED
        - USER_REGISTERED
        - USER_CHANGED_EMAIL
        - OTP_SENT
        - OTP_SEND_FAILED
        - LOGIN_SUCCESS
        - LOGIN_FAILED
        - LOGOUT
        - PASSWORD_RESET_REQUESTED
        - PASSWORD_RESET_COMPLETED
        - MFA_ENABLED
        - MFA_DISABLED
        - PASSWORD_CHANGED
        - PROFILE_UPDATED
        - UNAUTHORIZED_ACCESS
        - TOKEN_BLACKLISTED
        - RBAC_DENIED
        - ADMIN_CREATED_USER
        - ADMIN_USER_LOCK
        - ADMIN_USER_UNLOCK
        - ADMIN_FORCE_LOGOUT
        - ADMIN_ANNOUNCE_ALL
        - ROLE_CHANGED
        - ADMIN_ANNOUNCE_ROLE
        - ADMIN_DELETE_USER
        - DATA_EXPORT
...
    PublicAuditDto:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        event:
          $ref: '#/components/schemas/AuditEvent'
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time

    # -------------------------
    # Admin DTOs
    # -------------------------
...
    CreateBackupDto:
      type: object
      properties:
        name:
          type: string
        storage:
          type: string
          enum: [local, s3, both]
      example:
        name: nightly
        storage: local
    BackupMetadata:
      type: object
      properties:
        filename:
          type: string
        location:
          type: string
          enum: [local, s3]
        size:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId
paths:
...
  /users/name/{name}:
    get:
      tags:
        - users
      summary: Find user by name
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
  /users/me:
    get:
      tags:
        - users
      summary: Get current logged-in user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
    patch:
      tags:
        - users
      summary: Update current user
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
    delete:
      tags:
        - users
      summary: Delete current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /users/search:
    get:
      tags:
        - users
      summary: Search users (Instructor/Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserDto'
  /users/search-instructors:
    get:
      tags:
        - users
      summary: Search instructors (Admin/Student)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated users (role=Instructor)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserDto'
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: User by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
  /users/autocomplete/{q}:
    get:
      tags:
        - users
      summary: Autocomplete users (name/email suggestions)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    name:
                      type: string
                    email:
                      type: string
  /users/{id}/unread/increment:
    post:
      tags:
        - users
      summary: Increment unread notification counter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '204':
          description: No content
  /users/{id}/unread/reset:
    post:
      tags:
        - users
      summary: Reset unread notification counter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '204':
          description: No content

  # -------------------------
  # Feedback
  # -------------------------
  /feedback:
    post:
      tags:
        - feedback
      summary: Submit feedback (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeedbackDto'
      responses:
        '201':
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicFeedbackDto'
  /feedback/admin:
    get:
      tags:
        - feedback
      summary: List feedback (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated feedback list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicFeedbackDto'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
                  limit:
                    type: integer

  # -------------------------
  # Audit Log
  # -------------------------
...
  /audit/user/{userId}:
    get:
      tags:
        - audit
      summary: List audit events for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Audit logs for user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicAuditDto'

  # -------------------------
  # Admin
  # -------------------------
...
  /backups:
    post:
      tags:
        - backups
      summary: Create a backup (admin)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBackupDto'
      responses:
        '201':
          description: Backup metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupMetadata'
    get:
      tags:
        - backups
      summary: List backups
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Backup list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BackupMetadata'
...
  /backups/{id}:
    delete:
      tags:
        - backups
      summary: Delete backup metadata + file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
  /backups/download/{filename}:
    get:
      tags:
        - backups
      summary: Download backup file (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...